requirements.txt
streamlit
pandas
altair
pillow
python-dotenv
requests
beautifulsoup4
lxml
torch
torchvision
transformers
sqlalchemy
